feat(render): route WW3D render calls through WW3DBackend interface
