+++ Code/dxvk_wrapper/d3dx9.h
@@ -2,4 +2,6 @@
 
 #include "dxvk_wrapper_compat.h"
 
+#ifdef _WIN32
 #include <dxvk/d3d9.h>
+#endif
