+++ Code/dxvk_wrapper/windows.h
@@ -2,4 +2,6 @@
 
 #include "dxvk_wrapper_compat.h"
 
+#ifdef _WIN32
 #include <dxvk/windows.h>
+#endif
